Move realizationPicker utils into separate utils file
